requirements.txt
Django
psycopg2-binary
djangorestframework
django-filter
pillow
djangorestframework-simplejwt
drf-nested-routers
pytest-django
django-cors-headers
PyQt6
channels
requests
openai
coinbase_commerce